+++ src/math.ts
@@ -505,3 +505,7 @@ export const rangeIntersection = (
 
   return null;
 };
+
+export const isValueInRange = (value: number, min: number, max: number) => {
+  return value >= min && value <= max;
+};